docker/compose/automations/alice-auth/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
bcrypt>=4.0,<5.0
PyJWT>=2.8,<3.0
